LayoutTests/fast/canvas/image-buffer-resource-limits-expected.txt

Check that GPU process ImageBuffer resource limits are enforced

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS () => checkLimits(limits) is true
Increasing canvas count to 500
PASS () => renderingModeCounts.accelerated is 500
PASS () => renderingModeCounts.unaccelerated is 0
PASS () => renderingModeCounts.noBackend is 0
Increasing canvas count to 5000
PASS () => renderingModeCounts.accelerated is 5000
PASS () => renderingModeCounts.unaccelerated is 0
PASS () => renderingModeCounts.noBackend is 0
Increasing canvas count to 5500
PASS () => renderingModeCounts.accelerated is 5000
PASS () => renderingModeCounts.unaccelerated is 500
PASS () => renderingModeCounts.noBackend is 0
PASS () => nonCanvasImageBufferRenderingMode == "Accelerated" is true
Increasing canvas count to 50000
PASS () => renderingModeCounts.accelerated is 5000
PASS () => renderingModeCounts.unaccelerated is 45000
PASS () => renderingModeCounts.noBackend is 0
Increasing canvas count to 50500
PASS () => renderingModeCounts.accelerated is 5000
PASS () => renderingModeCounts.unaccelerated is 45000
PASS () => renderingModeCounts.noBackend is 500
Decreasing canvas count to 500
Increasing canvas count to 600
PASS () => renderingModeCounts.accelerated is 600
PASS () => renderingModeCounts.unaccelerated is 0
PASS () => renderingModeCounts.noBackend is 0
PASS successfullyParsed is true

TEST COMPLETE
